# schema/openapi.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "Continue API",
    "description": "Continue API",
    "version": "1.0"
  },
  "paths": {
...
    "/unittest/forline": {
      "post": {
        "tags": [
          "unittest"
        ],
        "summary": "Forline",
        "description": "Write unit test for the function encapsulating the given line number.",
        "operationId": "forline_unittest_forline_post",
        "parameters": [
          {
            "required": false,
            "schema": {
              "title": "X-Vsc-Machine-Id",
              "type": "string",
              "default": "anonymous"
            },
            "name": "x-vsc-machine-id",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FilePosition"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompletionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/unittest/failingtest": {
      "post": {
        "tags": [
          "unittest"
        ],
        "summary": "Failingtest",
        "description": "Write a failing test for the function encapsulating the given line number.",
        "operationId": "failingtest_unittest_failingtest_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FailingTestBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompletionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CompletionResponse": {
        "title": "CompletionResponse",
        "required": [
          "completion"
        ],
        "type": "object",
        "properties": {
          "completion": {
            "title": "Completion",
            "type": "string"
          }
        }
      },
...
      "FailingTestBody": {
        "title": "FailingTestBody",
        "required": [
          "description",
          "fp"
        ],
        "type": "object",
        "properties": {
          "description": {
            "title": "Description",
            "type": "string"
          },
          "fp": {
            "$ref": "#/components/schemas/FilePosition"
          }
        },
        "description": "A failing test body."
      },
...
      "FilePosition": {
        "title": "FilePosition",
        "required": [
          "filecontents",
          "lineno"
        ],
        "type": "object",
        "properties": {
          "filecontents": {
            "title": "Filecontents",
            "type": "string"
          },
          "lineno": {
            "title": "Lineno",
            "type": "integer"
          }
        },
        "description": "A position in a file."
      },
...
      "HTTPValidationError": {
        "title": "HTTPValidationError",
        "type": "object",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            }
          }
        }
      },
...
      "ValidationError": {
        "title": "ValidationError",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "type": "object",
        "properties": {
          "loc": {
            "title": "Location",
            "type": "array",
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            }
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        }
      }
    }
  }
}